requirements.txt
-r Sublist3r/requirements.txt
pydantic
pwntools